moved some functionality not specific to the Graph class to the utils so they can be exposed
