+++ spock/utils.py
@@ -25,7 +25,63 @@ else:
     from typing import _GenericAlias
 
 from pathlib import Path
-from typing import Union
+from enum import EnumMeta
+from typing import List, Type, Union
+
+
+def _find_all_spock_classes(attr_class: Type):
+    """Within a spock class determine if there are any references to other spock classes
+
+    Args:
+        attr_class: a class with attrs attributes
+
+    Returns:
+        list of dependent spock classes
+
+    """
+    # Get the top level dict
+    dict_attr = attr.fields_dict(attr_class)
+    # Dependent classes
+    dep_classes = []
+    for k, v in dict_attr.items():
+        # Checks for direct spock/attrs instance
+        if _is_spock_instance(v.type):
+            dep_classes.append(v.type)
+        # Check for enum of spock/attrs instance
+        elif isinstance(v.type, EnumMeta) and _check_4_spock_iterable(v.type):
+            dep_classes.extend(_get_enum_classes(v.type))
+        # Check for List[@spock-class] -- needs to be checked against 3.6 typing.List as well
+        elif ((v.type is list) or (v.type is List)) and _is_spock_instance(
+                v.metadata["type"].__args__[0]
+        ):
+            dep_classes.append(v.metadata["type"].__args__[0])
+    return dep_classes
+
+
+def _check_4_spock_iterable(iter_obj: Union[tuple, list]):
+    """Checks if an iterable type contains a spock class
+
+    Args:
+        iter_obj: iterable type
+
+    Returns:
+        boolean if the iterable contains at least one spock class
+
+    """
+    return _check_iterable(iter_obj=iter_obj)
+
+
+def _get_enum_classes(enum_obj: EnumMeta):
+    """Checks if any of the values of an enum are spock classes and adds to a list
+
+    Args:
+        enum_obj: enum class
+
+    Returns:
+        list of enum values that are spock classes
+
+    """
+    return [v.value for v in enum_obj if _is_spock_instance(v.value)]
 
 
 def path_object_to_s3path(path: Path) -> str:
